Move split_stack method to util module
